tests/requirements.txt
pytest-django

